new approach, based on cutter, added some rules, instructed infra-structures
